% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boshROI-package.R
\docType{package}
\name{boshROI}
\alias{boshROI}
\alias{boshROI-package}
\title{boshROI}
\description{
A package to assist in the quantification and analysis of fluorescence micrographs.
}
\keyword{package}
