% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{newRoiFolder}
\alias{newRoiFolder}
\title{newRoiFolder}
\usage{
newRoiFolder(folder, framePeriod = 1, firstRoiBaseline = TRUE)
}
\arguments{
\item{folder}{name of a directory that contains ROI data (character).}

\item{framePeriod}{period of time between image frames (numeric).}

\item{firstRoiBasline}{Indicates whether the first ROI in the dataset should be used for background normalization (logical).}
}
\value{
an RoiFolder object containing calculated dF/F values for each ROI
}
\description{
Defines a new RoiFolder object.
}
\details{
Reads a data file (e.g. "Results.xls"), experiment file (e.g. "experiment.txt"),
and an ROI-set zip file (e.g. "RoiSet.zip") from a parent directory ("folder").
Returns an RoiFolder object containing calculated dF/F values for each ROI (region of interest)
at each timepoint.
}
