% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{plot_ROI_DFF}
\alias{plot_ROI_DFF}
\title{plot_ROI_DFF}
\usage{
plot_ROI_DFF(ROI, saveAs = "NULL")
}
\arguments{
\item{ROI}{an RoiFolder object containing calculated dF/F values for each ROI.}

\item{saveAs}{(optional) A PNG filename (e.g. "figure.png"); Saves plot as a PNG file.}
}
\value{
a plot of dF/F values for each ROI vs. time
}
\description{
Plots dF/F of each ROI as a function of time from data contained in an RoiFolder object.
If saveAs=[PNG filename], saves plot as a PNG file.
}
\seealso{
Other RoiFolder functions: \code{\link{newRoiFolder}},
  \code{\link{plot_ROI_RAW}}
}
