% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{roi_folder_check_files}
\alias{roi_folder_check_files}
\title{Ensure an ROI analysis folder has the needed files.}
\usage{
roi_folder_check_files(folder)
}
\arguments{
\item{folder}{path to ROI folder}
}
\description{
This function looks for RoiSet.zip, experiment.txt, and
Results.xls. If one of these is missing, the program ends.
}
\examples{
roi_folder_check_files("/path/to/ROI/folder/")
}
