% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{roi_folder_process}
\alias{roi_folder_process}
\title{Automatically process a time series ROI folder.}
\usage{
roi_folder_process(folder)
}
\arguments{
\item{folder}{path to the time series folder.}
}
\value{
True if it worked, False if error.
}
\description{
This will create extra dF/F XLS documents, generate
figures, etc. The folder must contain RoiSet.zip,
experiment.txt, and Results.xls. Most likely it will
also contain a list of numerous TIF files time-stamped
with filenames equal to epoch seconds.
}
\examples{
roi_folder_process("/path/to/ROI/folder/")
}
