% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{roi_read_results}
\alias{roi_read_results}
\title{Load data from an ImageJ Results.xls}
\usage{
roi_read_results(results_file, timePeriod = 1)
}
\arguments{
\item{results_file}{path to results TSV file.}

\item{timePeriod}{time between captures. Leave =1 to frame number}
}
\value{
data table of results
}
\description{
In ImageJ the results of a multi-measure action are
usually saved as "Results.xls" (actually a TSV file).
This function parses that file and returns a table.
}
\examples{
roi_folder_process("/path/to/ROI/folder/Results.xls")
}
